Fix SRID issue when transformating tgeompoint <-> tgeogpoint (#440)
* Improve MEOS examples

* Add CodeQL workflow

* Fix SRID when transforming tgeogpoint <-> tgeompoint

* Fix CodeQL issue

+++ mobilitydb/src/point/tpoint_analyze.c
@@ -95,7 +95,7 @@
 /**
  * @brief The SD factor restricts the side of the statistics histogram
  * based on the standard deviation of the extent of the data.
- * 
+ *
  * SDFACTOR is the number of standard deviations from the mean
  * the histogram will extend.
  */
@@ -486,11 +486,11 @@ static int
 gbox_ndims(const GBOX* gbox)
 {
   int dims = 2;
-  if ( FLAGS_GET_GEODETIC(gbox->flags) )
+  if (FLAGS_GET_GEODETIC(gbox->flags))
     return 3;
-  if ( FLAGS_GET_Z(gbox->flags) )
+  if (FLAGS_GET_Z(gbox->flags))
     dims++;
-  if ( FLAGS_GET_M(gbox->flags) )
+  if (FLAGS_GET_M(gbox->flags))
     dims++;
   return dims;
 }
@@ -524,7 +524,7 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
   int histogram_features = 0;     /* # rows that actually got counted in the histogram */
 
   ND_STATS *nd_stats;             /* Our histogram */
-  size_t  nd_stats_size;          /* Size to allocate */
+  size_t nd_stats_size;           /* Size to allocate */
 
   double total_width = 0;         /* # of bytes used by sample */
   double total_cell_count = 0;    /* # of cells in histogram affected by sample */
@@ -535,15 +535,15 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
 
   const ND_BOX **sample_boxes;    /* ND_BOXes for each of the sample features */
   ND_BOX sample_extent;           /* Extent of the raw sample */
-  int  histo_size[ND_DIMS];       /* histogram nrows, ncols, etc */
+  int histo_size[ND_DIMS];        /* histogram nrows, ncols, etc */
   ND_BOX histo_extent;            /* Spatial extent of the histogram */
   ND_BOX histo_extent_new;        /* Temporary variable */
-  int  histo_cells_target;        /* Number of cells we will shoot for, given the stats target */
-  int  histo_cells;               /* Number of cells in the histogram */
-  int  histo_cells_new = 1;       /* Temporary variable */
+  int histo_cells_target;         /* Number of cells we will shoot for, given the stats target */
+  int histo_cells;                /* Number of cells in the histogram */
+  int histo_cells_new = 1;        /* Temporary variable */
 
-  int   ndims = 2;                /* Dimensionality of the sample */
-  int   histo_ndims = 0;          /* Dimensionality of the histogram */
+  int ndims = 2;                  /* Dimensionality of the sample */
+  int histo_ndims = 0;            /* Dimensionality of the histogram */
   double sample_distribution[ND_DIMS]; /* How homogeneous is distribution of sample in each axis? */
 
   int stats_slot;            /* What slot is this data going into? (2D vs ND) */
@@ -578,7 +578,7 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
    *  o compute total features's box area (for avgFeatureArea)
    *  o sum features box coordinates (for standard deviation)
    */
-  for ( i = 0; i < sample_rows; i++ )
+  for (i = 0; i < sample_rows; i++)
   {
     STBox box;
     GBOX gbox;
@@ -588,7 +588,7 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
     Datum datum = fetchfunc(stats, i, &is_null);
 
     /* Skip all NULLs. */
-    if ( is_null )
+    if (is_null)
     {
       null_cnt++;
       continue;
@@ -623,11 +623,11 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
     stbox_set_gbox(&box, &gbox);
 
     /* If we're in 2D mode, zero out the higher dimensions for "safety" */
-    if ( mode == 2 )
+    if (mode == 2)
       gbox.zmin = gbox.zmax = gbox.mmin = gbox.mmax = 0.0;
 
     /* Check bounds for validity (finite and not NaN) */
-    if ( ! gbox_is_valid(&gbox) )
+    if (! gbox_is_valid(&gbox))
     {
       continue;
     }
@@ -636,7 +636,7 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
      * In N-D mode, set the ndims to the maximum dimensionality found
      * in the sample. Otherwise, leave at ndims == 2.
      */
-    if ( mode != 2 )
+    if (mode != 2)
       ndims = Max(gbox_ndims(&gbox), ndims);
 
     /* Convert gbox to n-d box */
@@ -647,14 +647,14 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
     sample_boxes[notnull_cnt] = nd_box;
 
     /* Initialize sample extent before merging first entry */
-    if ( ! notnull_cnt )
+    if (! notnull_cnt)
       nd_box_init_bounds(&sample_extent);
 
     /* Add current sample to overall sample extent */
     nd_box_merge(nd_box, &sample_extent);
 
     /* Add bounds coordinates to sums for stddev calculation */
-    for ( d = 0; d < ndims; d++ )
+    for (d = 0; d < ndims; d++)
     {
       sum.min[d] += nd_box->min[d];
       sum.max[d] += nd_box->max[d];
@@ -675,7 +675,8 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
    * we have an average of 5 features for each cell so the histogram isn't
    * so sparse.
    */
-  histo_cells_target = (int)pow((double)(stats->attr->attstattarget), (double)ndims);
+  histo_cells_target = (int) pow((double) (stats->attr->attstattarget),
+    (double) ndims);
   histo_cells_target = Min(histo_cells_target, ndims * 10000);
   histo_cells_target = Min(histo_cells_target, (int)(total_rows/5));
 
@@ -683,14 +684,14 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
    * Second scan:
    *  o compute standard deviation
    */
-  for ( d = 0; d < ndims; d++ )
+  for (d = 0; d < ndims; d++)
   {
     /* Calculate average bounds values */
     avg.min[d] = sum.min[d] / notnull_cnt;
     avg.max[d] = sum.max[d] / notnull_cnt;
 
     /* Calculate standard deviation for this dimension bounds */
-    for ( i = 0; i < notnull_cnt; i++ )
+    for (i = 0; i < notnull_cnt; i++)
     {
       const ND_BOX *ndb = sample_boxes[i];
       stddev.min[d] += (ndb->min[d] - avg.min[d]) * (ndb->min[d] - avg.min[d]);
@@ -710,11 +711,11 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
    *   o compute new histogram box
    */
   nd_box_init_bounds(&histo_extent_new);
-  for ( i = 0; i < notnull_cnt; i++ )
+  for (i = 0; i < notnull_cnt; i++)
   {
     const ND_BOX *ndb = sample_boxes[i];
     /* Skip any hard deviants (boxes entirely outside our histo_extent */
-    if ( ! nd_box_intersects(&histo_extent, ndb, ndims) )
+    if (! nd_box_intersects(&histo_extent, ndb, ndims))
     {
       pfree((void *) sample_boxes[i]);
       sample_boxes[i] = NULL;
@@ -743,7 +744,7 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
    * homogeneous places).
    */
   nd_box_array_distribution(sample_boxes, notnull_cnt, &histo_extent, ndims,
-                sample_distribution);
+    sample_distribution);
 
   /*
    * The sample_distribution array now tells us how spread out the
@@ -759,21 +760,21 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
    * that have some interesting differences in data distribution.
    * Here we count up the number of interesting dimensions
    */
-  for ( d = 0; d < ndims; d++ )
+  for (d = 0; d < ndims; d++)
   {
-    if ( sample_distribution[d] > 0 )
+    if (sample_distribution[d] > 0)
       histo_ndims++;
   }
 
-  if ( histo_ndims == 0 )
+  if (histo_ndims == 0)
   {
     /* Special case: all our dimensions had low variability! */
     /* We just divide the cells up evenly */
     histo_cells_new = 1;
-    for ( d = 0; d < ndims; d++ )
+    for (d = 0; d < ndims; d++)
     {
       histo_size[d] = (int)pow((double)histo_cells_target, 1/(double)ndims);
-      if ( ! histo_size[d] )
+      if (! histo_size[d])
         histo_size[d] = 1;
       histo_cells_new *= histo_size[d];
     }
@@ -788,24 +789,26 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
     double total_distribution;       /* Total of sample_distribution */
     total_distribution = total_double(sample_distribution, ndims); /* First get the total */
     histo_cells_new = 1; /* For the number of cells in the final histogram */
-    for ( d = 0; d < ndims; d++ )
+    for (d = 0; d < ndims; d++)
     {
-      if ( sample_distribution[d] == 0 ) /* Uninteresting dimensions don't get any room */
+      if (sample_distribution[d] == 0) /* Uninteresting dimensions don't get any room */
       {
         histo_size[d] = 1;
       }
       else /* Interesting dimension */
       {
         /* How does this dims variability compare to the total? */
-        float edge_ratio = (float) sample_distribution[d] / (float) total_distribution;
+        float edge_ratio = (float) sample_distribution[d] /
+          (float) total_distribution;
         /*
          * Scale the target cells number by the # of dims and ratio,
          * then take the appropriate root to get the estimated number of cells
          * on this axis (eg, pow(0.5) for 2d, pow(0.333) for 3d, pow(0.25) for 4d)
         */
-        histo_size[d] = (int)pow(histo_cells_target * histo_ndims * edge_ratio, 1/(double)histo_ndims);
+        histo_size[d] = (int) pow((double) histo_cells_target * histo_ndims *
+          edge_ratio, 1 / (double) histo_ndims);
         /* If something goes awry, just give this dim one slot */
-        if ( ! histo_size[d] )
+        if (! histo_size[d])
           histo_size[d] = 1;
       }
       histo_cells_new *= histo_size[d];
@@ -831,7 +834,7 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
   nd_stats->table_features = (float4) total_rows;
   nd_stats->not_null_features = (float4) notnull_cnt;
   /* Copy in the histogram dimensions */
-  for ( d = 0; d < ndims; d++ )
+  for (d = 0; d < ndims; d++)
     nd_stats->size[d] = (float4) histo_size[d];
 
   /*
@@ -847,7 +850,7 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
    * histogram feature count.
    *
    */
-  for ( i = 0; i < notnull_cnt; i++ )
+  for (i = 0; i < notnull_cnt; i++)
   {
     const ND_BOX *nd_box;
     ND_IBOX nd_ibox;
@@ -858,7 +861,7 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
     double cellsize[ND_DIMS] = {0.0, 0.0, 0.0, 0.0};
 
     nd_box = sample_boxes[i];
-    if ( ! nd_box ) continue; /* Skip Null'ed out hard deviants */
+    if (! nd_box) continue; /* Skip Null'ed out hard deviants */
 
     /* Give backend a chance of interrupting us */
     vacuum_delay_point();
@@ -867,7 +870,7 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
     nd_box_overlap(nd_stats, nd_box, &nd_ibox);
     memset(at, 0, sizeof(int) * ND_DIMS);
 
-    for ( d = 0; d < nd_stats->ndims; d++ )
+    for (d = 0; d < nd_stats->ndims; d++)
     {
       /* Initialize the starting values */
       at[d] = nd_ibox.min[d];
@@ -885,7 +888,7 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
       ND_BOX nd_cell = { {0.0, 0.0, 0.0, 0.0}, {0.0, 0.0, 0.0, 0.0} };
       double ratio;
       /* Create a box for this histogram cell */
-      for ( d = 0; d < nd_stats->ndims; d++ )
+      for (d = 0; d < nd_stats->ndims; d++)
       {
         nd_cell.min[d] = (float4) (min[d] + (at[d]+0) * cellsize[d]);
         nd_cell.max[d] = (float4) (min[d] + (at[d]+1) * cellsize[d]);
@@ -909,13 +912,13 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
   }
 
   /* Free memory */
-  for ( i = 0; i < notnull_cnt; i++ )
-    if ( sample_boxes[i] != NULL )
+  for (i = 0; i < notnull_cnt; i++)
+    if (sample_boxes[i] != NULL)
       pfree((void *) sample_boxes[i]);
   pfree(sample_boxes);
 
   /* Error out if we got no sample information */
-  if ( ! histogram_features )
+  if (! histogram_features)
   {
     elog(NOTICE, " no features lie in the stats histogram, invalid stats");
     stats->stats_valid = false;
@@ -927,7 +930,7 @@ gserialized_compute_stats(VacAttrStats *stats, AnalyzeAttrFetchFunc fetchfunc,
   nd_stats->cells_covered = (float4) total_cell_count;
 
   /* Put this histogram data into the right slot/kind */
-  if ( mode == 2 )
+  if (mode == 2)
   {
     stats_slot = STATISTIC_SLOT_2D;
     stats_kind = STATISTIC_KIND_2D;
