+++ postgis/liblwgeom/ptarray.c
@@ -232,11 +232,12 @@ ptarray_append_ptarray(POINTARRAY *pa1, POINTARRAY *pa2, double gap_tolerance)
 	{
 		pa1->maxpoints = ncap > pa1->maxpoints*2 ?
 		                 ncap : pa1->maxpoints*2;
-		pa1->serialized_pointlist = lwrealloc(pa1->serialized_pointlist, ptsize * pa1->maxpoints);
+		pa1->serialized_pointlist = lwrealloc(pa1->serialized_pointlist,
+      (size_t) ptsize * pa1->maxpoints);
 	}
 
 	memcpy(getPoint_internal(pa1, pa1->npoints),
-	       getPoint_internal(pa2, poff), ptsize * npoints);
+	       getPoint_internal(pa2, poff), (size_t) ptsize * npoints);
 
 	pa1->npoints = ncap;
